remove pollUtils.js
Put all poll utils in `/polls/utils.js`
